+++ src/client/editor/tools/FocusTool.tsx
@@ -1,6 +1,8 @@
 import { memo } from 'react';
 import SymbolTool from '../SymbolTool';
-import { instance } from '@logic-pad/core/data/symbols/focusSymbol';
+import FocusSymbolData, {
+  instance,
+} from '@logic-pad/core/data/symbols/focusSymbol';
 import FocusSymbol from '../../symbols/FocusSymbol';
 
 const sample = instance;
@@ -8,12 +10,22 @@ const sample = instance;
 export default memo(function FocusTool() {
   return (
     <SymbolTool
-      name="Focus"
+      name="Focus/Dead End"
       order={202}
       hotkey="extraSymbols-1"
       defaultHidden={true}
       sample={sample}
       component={FocusSymbol}
+      onNewSymbol={(symbol, grid) => {
+        const existing = grid.findSymbol(
+          sym =>
+            sym.x === symbol.x && sym.y === symbol.y && sym.id !== symbol.id
+        );
+        if (existing) {
+          return (symbol as FocusSymbolData).withDeadEnd(true);
+        }
+        return symbol;
+      }}
     />
   );
 });