+++ src/client/editor/tools/HiddenTool.tsx
@@ -1,8 +1,6 @@
 import { memo } from 'react';
 import SymbolTool from '../SymbolTool';
-import HiddenSymbolData, {
-  instance,
-} from '@logic-pad/core/data/symbols/hiddenSymbol';
+import { instance } from '@logic-pad/core/data/symbols/hiddenSymbol';
 import HiddenSymbol from '../../symbols/HiddenSymbol';
 
 const sample = instance;
@@ -16,9 +14,6 @@ export default memo(function HiddenTool() {
       defaultHidden={true}
       sample={sample}
       component={HiddenSymbol}
-      onNewSymbol={symbol => {
-        return symbol as HiddenSymbolData;
-      }}
     />
   );
 });